Fixing new db
