+++ api/db/base.py
@@ -1,0 +1,3 @@
+from db.database import Base
+from db.models.company import Company
+from db.models.netassetvalue import NetAssetValue
